Move shell context menu stuff to helper class
